refactor: renaming segmented-button to segment and segmented-button-group to segmented-button
